Begin refactor of tests
